SmartDrain/requirements.txt
absl-py==2.1.0
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.8.1
astunparse==1.6.3
automium==0.2.6
automium_web==0.1.1
awscli==1.35.19
billiard==4.2.1
blinker==1.9.0
boto3==1.35.53
botocore==1.35.53
cachetools==5.5.2
celery==5.5.3
certifi==2024.7.4
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
contourpy==1.3.0
cron-descriptor==1.4.5
cycler==0.12.1
Django==5.0.7
django-celery-beat==2.8.1
django-cors-headers==4.4.0
django-filter==25.1
django-request-logging==0.7.5
django-timezone-field==7.1
djangorestframework==3.15.2
docker==7.1.0
docstring_parser==0.16
docutils==0.16
et-xmlfile==1.1.0
filelock==3.16.1
Flask==3.1.1
flatbuffers==24.3.25
fonttools==4.54.1
fsspec==2024.10.0
gast==0.6.0
google-api-core==2.24.2
google-auth==2.40.2
google-cloud-aiplatform==1.71.1
google-cloud-bigquery==3.33.0
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.2
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpc-google-iam-v1==0.14.2
grpcio==1.64.1
grpcio-status==1.71.0
h5py==3.11.0
idna==3.7
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.5.1
keras==3.4.1
keyring==8.7
keyrings.alt==5.0.2
kiwisolver==1.4.7
kombu==5.5.4
legacy==0.1.7
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
ml-dtypes==0.4.0
more-itertools==10.5.0
mpmath==1.3.0
mysqlclient==2.2.4
namex==0.0.8
networkx==3.4.2
numpy==1.26.4
opencv-python==4.10.0.84
openpyxl==3.1.5
opt-einsum==3.3.0
optree==0.12.1
packaging==24.1
pandas==2.2.2
pillow==11.0.0
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==4.25.3
psutil==6.1.0
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.18.0
pyparsing==3.2.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
pytz==2024.1
pywin32==308
pywin32-ctypes==0.2.3
PyYAML==6.0.2
requests==2.32.3
rich==13.7.1
rsa==4.7.2
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.12
s3transfer==0.10.3
scikit-learn==1.4.2
scipy==1.14.1
seaborn==0.13.2
sequential==1.0.0
setuptools==70.3.0
shapely==2.1.1
six==1.16.0
sqlparse==0.5.1
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-intel==2.17.0
termcolor==2.4.0
threadpoolctl==3.6.0
torch==2.5.1
torchvision==0.20.1
tqdm==4.66.6
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
ultralytics==8.3.27
ultralytics-thop==2.0.10
urllib3==2.2.2
vertexai==1.71.1
vine==5.1.0
voluptuous==0.15.2
wcwidth==0.2.13
Werkzeug==3.1.3
wheel==0.43.0
wrapt==1.16.0
yolo==0.3.1
